Former-commit-id: 74ab64b6d735ee2497820e8fa18abef2aa885007

+++ Sources/Tools/CommandLine/CommandLineParser.cpp
@@ -162,7 +162,7 @@ DAVA::int32 CommandLineParser::GetCommandPosition(const DAVA::String& command)
     {
         if (command == commandLine[i])
         {
-            position = i;
+            position = static_cast<int32>(i);
             break;
         }
     }
